Edit component's text indent rules
