+++ src/router/index.js
@@ -1,6 +1,6 @@
-import Vue from 'vue'
-import Router from 'vue-router'
-import Front from '@/components/Front'
+import Vue from 'vue';
+import Router from 'vue-router';
+import Main from '@/components/Main.vue';
 
 Vue.use(Router)
 
@@ -8,8 +8,8 @@ export default new Router({
   routes: [
     {
       path: '/',
-      name: 'Front',
-      component: Front
+      name: 'main',
+      component: Main
     }
   ]
 })
